Shared Library Restructure.
 Move original components made my habomane and me to a legacy folder.
 Update imports for components in the shared library.
 Update exports for components in the shared library.
