+++ libs/shared/src/components/atoms/index.ts
@@ -1,5 +1,5 @@
-export * from './button/button.component';
+// export * from './button/button.component';
 export * from './card/card.component';
 export * from './chip/chip.component';
-export * from './divider/divider.component';
+// export * from './divider/divider.component';
 export * from './icon/icon.component';
