+++ libs/shared/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './atoms/index';
+export * from './molecules/index';
+export * from './legacy/index';
