--- libs/shared/src/components/inputs/index.ts
@@ -1,1 +1,0 @@
-export * from './text-input/text-input.component';
