--- libs/shared/src/components/inputs/text-input/text-input.component.html
@@ -1,4 +1,0 @@
-<div class="form-group">
-   <label for="input">Input:</label>
-   <input type="text" class="form-control" id="input" />
-</div>
