--- libs/shared/src/components/inputs/text-input/text-input.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TextInputComponent } from './text-input.component';
-
-describe('TextInputComponent', () => {
-   let component: TextInputComponent;
-   let fixture: ComponentFixture<TextInputComponent>;
-
-   beforeEach(async () => {
-      await TestBed.configureTestingModule({
-         imports: [TextInputComponent]
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(TextInputComponent);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-   });
-
-   it('should create', () => {
-      expect(component).toBeTruthy();
-   });
-});
