+++ libs/shared/src/components/legacy/advisory-card/advisory-card.component.html
