+++ libs/shared/src/components/legacy/advisory-card/advisory-card.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AdvisoryCard } from '../../models';
+import { DropdownComponent } from '../dropdown/dropdown.component';
 import { ButtonComponent } from '../button/button.component';
 import { CardHeaderComponent } from '../card-header/card-header.component';
-import { DropdownComponent } from '../dropdown/dropdown.component';
+import { AdvisoryCard } from '../../../models/index';
 
 @Component({
    selector: 'lib-advisory-card',
