+++ libs/shared/src/components/legacy/divider/divider.component.scss
