+++ libs/shared/src/components/legacy/dropdown/dropdown.component.html
