+++ libs/shared/src/components/legacy/dropdown/dropdown.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { Dropdown } from '../../models';
+import { Dropdown } from '../../../models';
 
 @Component({
    selector: 'lib-dropdown',
