+++ libs/shared/src/components/legacy/hero/hero.component.scss
