+++ libs/shared/src/components/legacy/index.ts
@@ -1,0 +1,8 @@
+export * from './advisory-card/advisory-card.component';
+export * from './loading-skeleton/loading-skeleton.component';
+export * from './tab/tab.component';
+export * from './hero/hero.component';
+export * from './dropdown/dropdown.component';
+export * from './divider/divider.component';
+export * from './card-header/card-header.component';
+export * from './button/button.component'
