+++ libs/shared/src/components/legacy/loading-skeleton/loading-skeleton.component.html
