+++ libs/shared/src/components/legacy/loading-skeleton/loading-skeleton.component.scss
@@ -1,5 +1,5 @@
 //* Component Variables */
-@import '../../../../assets/main.scss';
+@import '../../../../../assets/main.scss';
 
 //* Component Variables */
 $button-font: 'General Sans', sans-serif;
