+++ libs/shared/src/components/legacy/tab/tab.component.scss
