+++ libs/shared/src/index.ts
@@ -1,12 +1,5 @@
 // Exporting Components
-export * from './components/advisory-card/advisory-card.component';
-export * from './components/button/button.component';
-export * from './components/card-header/card-header.component';
-export * from './components/hero/hero.component';
-export * from './components/tab/tab.component';
-export * from './components/dropdown/dropdown.component';
-export * from './components/loading-skeleton/loading-skeleton.component';
-export * from './components/divider/divider.component';
+export * from './components/index'
 
 // Exporting Mock, Types, and Interfaces
 export * from './mocks';
